Rename *Engine
Rename *engine to *manager and collect them all in the directory
controllers.

The classes were renamed because the exact same name was used in
a different project. They were also moved since the base class was in
the wrong file.
